==> notebooks/1.Extracting feature Vectors from Images.ipynb <==
# Importing Libraries

import numpy as np
from numpy import array
import pandas as pd
import matplotlib.pyplot as plt 
import string
import os
from PIL import Image
import glob
from pickle import dump, load
from tqdm import tqdm_notebook as tqdm
from time import time
from keras.preprocessing import sequence
from keras.models import Sequential
from keras.layers import LSTM, Embedding, TimeDistributed, Dense, RepeatVector,\
                         Activation, Flatten, Reshape, concatenate, Dropout, BatchNormalization
from keras.optimizers import Adam, RMSprop
from keras.layers.wrappers import Bidirectional
from keras.layers.merge import add
from keras.applications.inception_v3 import InceptionV3
from keras.preprocessing import image
from keras.models import Model
from keras import Input, layers
from keras import optimizers
from keras.applications.inception_v3 import preprocess_input
from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
from keras.utils import to_categorical

from google.colab import drive
drive.mount('/content/drive')

try:
    import dill as pickle
except ImportError:
    import pickle

# Loading Documents

# load doc into memory
def load_doc(filename): 
    file = open(filename, 'r', encoding='utf8') 
    text = file.read() 
    file.close()
    return text

#filename = "./drive/My Drive/Flickr8K_Text/Flickr8k.token.txt"
#filename = "./drive/My Drive/Project/Theme_data_text/themedata.token.txt"
filename = "./drive/My Drive/Project/urdu/urdu_token.txt"
# load descriptions
doc = load_doc(filename)
print(doc[:300])

def load_descriptions(doc):
    mapping = dict()
    
    for line in doc.split('\n'):
        
        if len(line) < 2:
            continue
        tokens = line.split()
        
        image_id, image_desc = tokens[0], tokens[1:]
        
        image_id = image_id.split('.')[0]
        
        image_desc = ' '.join(image_desc)
        
        if image_id not in mapping:
            mapping[image_id] = list()
        
        mapping[image_id].append(image_desc)
    return mapping

descriptions = load_descriptions(doc)
print('Loaded: %d ' % len(descriptions))

list(descriptions.items())[:1]

from nltk.tokenize import RegexpTokenizer
tokenizer = RegexpTokenizer(r'\w+')

def clean_descriptions(descriptions):
    
    table = str.maketrans('', '', string.punctuation)
    for key, desc_list in descriptions.items():
        for i in range(len(desc_list)):
            desc = desc_list[i]
            
            desc = tokenizer.tokenize(desc)
            
            desc = [word.lower() for word in desc]
            
            desc = [word for word in desc if len(word)>1]
            
            desc_list[i] =  ' '.join(desc)

clean_descriptions(descriptions)

descriptions['1015118661_980735411b']

# convert the loaded descriptions into a vocabulary of words
def getVocabulary(descriptions):
    # build a list of all description strings
    all_desc = set()
    for key in descriptions.keys():
        [all_desc.update(tokenizer.tokenize(d)) for d in descriptions[key]]
    return all_desc

# summarize vocabulary
vocabulary = getVocabulary(descriptions)
print('Original Vocabulary Size: %d' % len(vocabulary))

# save descriptions to file, one per line
def save_descriptions(descriptions, filename):
    lines = list()
    for key, desc_list in descriptions.items():
        for desc in desc_list:
            lines.append(key + ' ' + desc)
    data = '\n'.join(lines)
    file = open(filename, 'w')
    file.write(data)
    file.close()

#save_descriptions(descriptions, 'descriptions.txt')
save_descriptions(descriptions, 'urdu_descriptions.txt')
print("Desciption.txt created")

# Preparing Data for training

# load Train Data
def load_set(filename):
    doc = load_doc(filename)
    dataset = list()
    
    for line in doc.split('\n'):
        
        if len(line) < 1:
            continue
 
        dataset.append(line)
    return (dataset)
 
#filename = './drive/My Drive/Flickr8K_Text/Flickr_8k.trainImages.txt'
#filename = './drive/My Drive/Project/Theme_data_text/train.names.txt'
filename = './drive/My Drive/Project/urdu/train_urdu.txt'
train_names = load_set(filename)
print('Train size: %d' % len(train_names))

#Directory_path = './drive/My Drive/Flickr8K_Data/'
Directory_path = './drive/My Drive/Project/Theme_dataset/'
train_img= [Directory_path+i  for i in train_names]

# Loading test data, 
def load_set(filename):
    doc = load_doc(filename)
    dataset = list()
    
    for line in doc.split('\n'):
        
        if len(line) < 1:
            continue
 
        dataset.append(line)
    return (dataset)
 
#filename = './drive/My Drive/Flickr8K_Text/Flickr_8k.testImages.txt'
#filename = './drive/My Drive/Project/Theme_data_text/test.names.txt'
filename = './drive/My Drive/Project/urdu/test_urdu.txt'
test_names = load_set(filename)
print('Test Size: %d' % len(test_names))

#Directory_path = './drive/My Drive/Flickr8K_Data/'
Directory_path = './drive/My Drive/Project/Theme_dataset/'
test_img= [Directory_path+i  for i in test_names]

# we r extracting train data captions such that, each captions starts with 'startseq' and ends with 'endseq
def load_clean_descriptions(filename, dataset):
    # load document
    dataset = [i.split('.')[0] for i in dataset]
    doc = load_doc(filename)
    descriptions = dict()
    for line in doc.split('\n'):
        
        tokens = line.split()
        
        image_id, image_desc = tokens[0], tokens[1:]
        
        if image_id in dataset:
            
            if image_id not in descriptions:
                descriptions[image_id] = list()
            
            desc = 'startseq ' + ' '.join(image_desc) + ' endseq'
            
            descriptions[image_id].append(desc)
    return descriptions

# descriptions
train_descriptions = load_clean_descriptions('urdu_descriptions.txt', train_names)

print('Descriptions: train=%d' % len(train_descriptions))

# Save train_descriptions
#with open("train_descriptions.pkl", "wb") as encoded_pickle:
    #pickle.dump(train_descriptions, encoded_pickle)
#with open("theme_train_descriptions.pkl", "wb") as encoded_pickle:
with open("urdu_train_descriptions.pkl", "wb") as encoded_pickle:
    pickle.dump(train_descriptions, encoded_pickle)

# Extracting feature Vectors from Images

def preprocess(image_path):
    # Convert all the images to size 299x299 as expected by the inception v3 model
    img = image.load_img(image_path, target_size=(299, 299))
    
    x = image.img_to_array(img)
    # Add one more dimension
    x = np.expand_dims(x, axis=0)
    
    x = preprocess_input(x)
    return x

# Load the inception v3 model
model = InceptionV3(weights='imagenet')

model.summary()

# Create a new model, by removing the last layer (output layer) from the inception v3
model_new = Model(model.input, model.layers[-2].output)

# Function to encode a given image into a vector of size (2048, )
def encode(image):
    image = preprocess(image) 
    v = model_new.predict(image) 
    v = np.reshape(v, v.shape[1]) 
    return v

# Encoding all the Training Dataset to a 2048 Dimension Vectors, Run this once
start = time()
k=0
encoded_train = {}
for img in tqdm(train_img):
 
    encoded_train[img[len(Directory_path):]] = encode(img)
print("Time taken in seconds =", time()-start)

# Encoding all the Test Dataset to a 2048 Dimension Vectors, Run this once
start = time()
encoded_test = {}
for img in tqdm(test_img):
    encoded_test[img[len(Directory_path):]] = encode(img)
print("Time taken in seconds =", time()-start)

# Save Train features
#with open("encoded_train_images.pkl", "wb") as encoded_pickle:
    #pickle.dump(encoded_train, encoded_pickle)
#with open("theme_encoded_train_images.pkl", "wb") as encoded_pickle:
with open("urdu_encoded_train_images.pkl", "wb") as encoded_pickle:
    pickle.dump(encoded_train, encoded_pickle)

# Save test features 

from pickle import dump, load
#with open("encoded_test_images.pkl", "wb") as encoded_pickle:
 #   pickle.dump(encoded_test, encoded_pickle)

#with open("theme_encoded_test_images.pkl", "wb") as encoded_pickle:
with open("urdu_encoded_test_images.pkl", "wb") as encoded_pickle:
    pickle.dump(encoded_test, encoded_pickle)

 #   End Of this notebook

for key, c in encoded_train.items():
    print(key)

train_img[0][len(Directory_path):]

len(encoded_train)
