==> notebooks/4. Caption Generator .ipynb <==
import numpy as np
from numpy import array
import pandas as pd
import matplotlib.pyplot as plt 
import string
import os
from PIL import Image
import glob
from pickle import dump, load
from tqdm import tqdm_notebook as tqdm
from time import time
from keras.preprocessing import sequence
from keras.models import Sequential
from keras.layers import LSTM, Embedding, TimeDistributed, Dense, RepeatVector,\
                         Activation, Flatten, Reshape, concatenate, Dropout, BatchNormalization
import h5py
from keras.optimizers import Adam, RMSprop
from keras.layers.wrappers import Bidirectional
from keras.layers.merge import add
from keras.applications.inception_v3 import InceptionV3
from keras.preprocessing import image
from keras.models import Model
from keras import Input, layers
from keras import optimizers
from keras.applications.inception_v3 import preprocess_input
from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
from keras.utils import to_categorical
import nltk

try:
    import dill as pickle
except ImportError:
    import pickle

from google.colab import drive
drive.mount('/content/drive')

# Loading Caption Generator Model

#model = load(open('./drive/My Drive/Project/theme_Caption_generator_Model.pkl', 'rb'))
model = load(open('./drive/My Drive/Project/urdu/urdu_Caption_generator_Model.pkl', 'rb'))

model.summary()

#images = './drive/My Drive/Flickr8K_Data/'
images = './drive/My Drive/Project/Theme_dataset/'

#with open("./drive/My Drive/Project/theme_encoded_test_images.pkl", "rb") as encoded_pickle:
#with open("./drive/My Drive/Project/urdu/urdu_encoded_test_images.pkl", "rb") as encoded_pickle:
with open("./drive/My Drive/Project/urdu/urdu_encoded_test_images.pkl", "rb") as encoded_pickle:
    encoding_test = load(encoded_pickle)

# Loading wordtoix
#wordtoix = load(open("./drive/My Drive/Project/theme_wordtoix.pkl", "rb"))
wordtoix = load(open("./drive/My Drive/Project/urdu/urdu_wordtoix.pkl", "rb"))
print('Wordtoix Length = %d' % len(wordtoix))

ixtoword = {}
for word, i in wordtoix.items():
    ixtoword[i] = word

Directory_path = "./drive/My Drive/Project/Theme_dataset/"

def greedySearch(photo):
    in_text = 'startseq'
    for i in range(max_length):
        sequence = [wordtoix[w] for w in in_text.split() if w in wordtoix]
        sequence = pad_sequences([sequence], maxlen=max_length)
        yhat = model.predict([photo,sequence], verbose=0)
        yhat = np.argmax(yhat)
        word = ixtoword[yhat]
        in_text += ' ' + word
        if word == 'endseq':
            break
    final = in_text.split()
    final = final[1:-1]
    final = ' '.join(final)
    return final

max_length = 35

z =0
results=[]
while (z<len(encoding_test)):
  pic = list(encoding_test.keys())[z]
  image = encoding_test[pic].reshape((1,2048))
  #image = encoding_test[pic].reshape((1,4096))
  results.append(greedySearch(image))
  x=plt.imread(Directory_path+pic)
  plt.axis('off')
  plt.imshow(x)
  plt.show()
  print(greedySearch(image))
  z+=1

# Bleu Score Calculation

filepath = './drive/My Drive/Project/urdu/test_captions.txt'
ground_captions=[]
with open(filepath, encoding='utf-8') as fp:
    names = fp.readline()
    while names:
        names=names.replace("\n","")
        ground_captions.append(names)
        names = fp.readline()

z=0
results=[]
Bleuscore=[]
while (z<len(encoding_test)):
  pic = list(encoding_test.keys())[z]
  image = encoding_test[pic].reshape((1,2048))
  #image = encoding_test[pic].reshape((1,4096))
  results.append(greedySearch(image))

  score = nltk.translate.bleu_score.sentence_bleu([ground_captions[z*5],ground_captions[(z*5)+1],ground_captions[(z*5)+2],ground_captions[(z*5)+3],ground_captions[(z*5)+4]], results[z])
  print("image",z,score)
  Bleuscore.append(score)
  z+=1

print("Bleu Score of model is:", (sum(Bleuscore)*100)/len(encoding_test))

z =0
results=[]
pic = list(encoding_test.keys())[z]

pic

  image = encoding_test[pic]

image.shape
